pyproject.toml
[tool.poetry]
name = "repl-nix-smart-health-care"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_smart_health_care"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
sqlalchemy = "^2.0.29"
mysql-connector-python = "^8.3.0"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
